+++ backend/canisters/community/impl/src/updates/c2c_join_community.rs
@@ -156,6 +156,9 @@ pub(crate) fn join_community_impl(
     match result {
         AddResult::Success(_) => {
             state.data.former_members.remove(&args.user_id);
+            state
+                .data
+                .cache_migrations_if_former_member(args.user_id, &args.previous_user_ids);
         }
         AddResult::AlreadyInCommunity => {
             let member = state.data.members.get_by_user_id(&args.user_id).unwrap();